requirements.txt
# PDF Extraction & Chunking
PyMuPDF==1.23.8
transformers==4.38.1
torch==2.2.1
tqdm==4.66.2

# Embedding & Supabase Vector DB
sentence-transformers==2.5.1
numpy<2
supabase==2.3.5
gotrue==1.3.1
httpx==0.24.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9

# API & Backend (Future phases)
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
groq
